bot/requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
idna==3.2
loguru==0.5.3
multidict==5.1.0
pymongo==3.12.0
pytz==2021.1
PyYAML==5.4.1
typing-extensions==3.10.0.0
yarl==1.6.3